\name{APAFULL}
\alias{APAFULL}
\docType{data}
\title{
American Psychological Association dataset, full version
}
\description{
The American Psychological Association dataset includes 15449 ballots of the election of the president in 1980, 5738 of which are complete rankings, in which the candidates are ranked from most to least favorite.
}
\usage{data(APAFULL)}


\source{
Diaconis, P. (1988). Group representations in probability and statistics. Lecture Notes-Monograph Series, i-192., pag. 96.
}

\keyword{datasets}
