% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branches.R
\name{branches}
\alias{branches}
\title{Auxliary code recalled by other routines}
\usage{
branches(brR, cij, b, Po, ord, Pb, FULL = FALSE)
}
\arguments{
\item{brR}{Current processed branche of the BB algorithm}

\item{cij}{Combined input matrix}

\item{b}{Other inputs recalled by main functions}

\item{Po}{Other inputs recalled by main functions}

\item{ord}{Other inputs recalled by main functions}

\item{Pb}{Other inputs recalled by main functions}

\item{FULL}{Default FULL=FALSE. If FULL=TRUE, the searching is limited to the space of complete rankings. TRUE=TRUE if the function is called by BBFULL algorithm.}
}
\value{
a "list" containing the following components:
\tabular{lll}{
cR \tab  \tab ranking belonging to the branche\cr
pcR \tab       \tab penalty of the current ranking}
}
\description{
Branches discovery
}
\author{
Antonio D'Ambrosio \email{antdambr@unina.it}
}
