% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childclosint.R
\name{childclosint}
\alias{childclosint}
\title{Transform the vector into ranking for DECOR}
\usage{
childclosint(r)
}
\arguments{
\item{r}{mutated vector}
}
\value{
a valid ranking
}
\description{
Closest integer approach is used: the elements of x are rounded
to the closest integer. Then check if any solution exists outside 
of the bounds (and get it back inside the bounds randomly).
Finally repair the solution if repetitions exist.
}
\references{
Davendra, D., and Onwubolu, G. (2007). Enhanced differential evolution hybrid scatter search for discrete optimization. In Evolutionary Computation, 2007. CEC 2007. IEEE Congress on (pp. 1156-1162). IEEE.
}
\author{
Antonio D'Ambrosio \email{antdambr@unina.it} and Giulio Mazzeo \email{giuliomazzeo@gmail.com}
}
