% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childtie.R
\name{childtie}
\alias{childtie}
\title{Auxiliary function}
\usage{
childtie(r)
}
\arguments{
\item{r}{a candidate rank vector}
}
\value{
a ranking
}
\description{
Apply discretization and convert to rank
}
\author{
Antonio D'Ambrosio \email{antdambr@unina.it} and Giulio Mazzeo \email{giuliomazzeo@gmail.com}
}
