% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combincost.R
\name{combincost}
\alias{combincost}
\title{Auxiliary function called by DECORcore}
\usage{
combincost(ranking, cij, M)
}
\arguments{
\item{ranking}{the candidate ranking}

\item{cij}{combined input matrix}

\item{M}{number of judges}
}
\value{
a "list" containing the following components:
\tabular{lll}{
tp \tab  \tab tauX associated to the ranking\cr
cp \tab       \tab cost associated to the ranking}
}
\description{
Calculates the sum of distances between a candidate ranking and the data set of rankings
}
\author{
Antonio D'Ambrosio \email{antdambr@unina.it} and Giulio Mazzeo
}
