% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{combinpmatr_impl}
\alias{combinpmatr_impl}
\title{Combined input matrix (C++ implementation)}
\usage{
combinpmatr_impl(X, Wk = NULL)
}
\arguments{
\item{X}{A numeric matrix (N x M) where rows are judges and columns are objects}

\item{Wk}{Optional numeric vector of weights for each ranking}
}
\value{
The M x M combined input matrix
}
\description{
Combined input matrix (C++ implementation)
}
\keyword{internal}
