% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{ConsRank-defunct}
\alias{ConsRank-defunct}
\alias{EMCons}
\alias{QuickCons}
\alias{BBFULL}
\alias{FASTcons}
\alias{DECOR}
\alias{FASTDECOR}
\alias{labels}
\title{Defunct functions in ConsRank}
\usage{
EMCons(...)

QuickCons(...)

BBFULL(...)

FASTcons(...)

DECOR(...)

FASTDECOR(...)

labels(...)
}
\description{
These functions have been removed from ConsRank. Use the alternatives listed below.
}
\details{
\itemize{
  \item \code{EMCons}: Use \code{\link{consrank}} with \code{algorithm = "BB"}
  \item \code{QuickCons}: Use \code{\link{consrank}} with \code{algorithm = "quick"}
  \item \code{BBFULL}: Use \code{\link{consrank}} with \code{algorithm = "BB"} and \code{full = TRUE}
  \item \code{FASTcons}: Use \code{\link{consrank}} with \code{algorithm = "fast"}
  \item \code{DECOR}: Use \code{\link{consrank}} with \code{algorithm = "decor"}
  \item \code{FASTDECOR}: Use \code{\link{consrank}} with \code{algorithm = "fastdecor"}
  \item \code{labels}: Functionality merged into \code{\link{rank2order}}
}
}
\keyword{internal}
