% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{ConsRank-deprecated}
\alias{ConsRank-deprecated}
\title{Deprecated functions in ConsRank}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
 \item \code{\link{EMCons}};
 \item \code{\link{QuickCons}};
 \item \code{\link{BBFULL}};
 \item \code{\link{FASTcons}};
 \item \code{\link{DECOR}};
 \item \code{\link{FASTDECOR}};
 \item \code{\link{labels}};
 } 
 
 All these  functions are deprecated, and will
 be removed in the next release of this package.
 The functions still remain in the package for compatibility of ConsRank users
}
\seealso{
\code{\link{consrank}}

\code{\link{rank2order}}
}
