% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{crossover}
\alias{crossover}
\title{Apply the (binomial) crossover for DE algorithm}
\usage{
crossover(x, v, CR)
}
\arguments{
\item{x}{target ranking}

\item{v}{donor individual (mutaded x)}

\item{CR}{Crossover range}
}
\value{
modified ranking
}
\description{
Binomial crossover stipulates that crossover will occur on each 
of the D values in a solution whenever a randomly generated number
between 0 and 1 is within the CR range.
}
\author{
Antonio D'Ambrosio \email{antdambr@unina.it} and Giulio Mazzeo \email{giuliomazzeo@gmail.com}
}
