% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findconsensusBB.R
\name{findconsensusBB}
\alias{findconsensusBB}
\title{Find a first ranking candidate to be the median ranking}
\usage{
findconsensusBB(cij)
}
\arguments{
\item{cij}{combined input matrix}
}
\value{
candidate median ranking
}
\description{
Auxiliary function: it finds a first ranking candidate to be the consensus ranking by observing the Combined input matrix
}
\author{
Antonio D'Ambrosio \email{antdambr@unina.it}
}
