% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutaterand1.R
\name{mutaterand1}
\alias{mutaterand1}
\title{Mutation phase}
\usage{
mutaterand1(X, FF, i)
}
\arguments{
\item{X}{population matrix}

\item{FF}{scaling factor}

\item{i}{population index to be ignored}
}
\value{
the mutated vector
}
\description{
Creates mutation vector v based on the current population X. We use the
rand/1/bin system
}
\author{
Antonio D'Ambrosio \email{antdambr@unina.it} and Giulio Mazzeo \email{giuliomazzeo@gmail.com}
}
