% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order2rank.R
\name{order2rank}
\alias{order2rank}
\title{Given an ordering, it is transformed to a ranking}
\usage{
order2rank(X, TO = "{", TC = "}", NAs = "NA")
}
\arguments{
\item{X}{A ordering or a matrix containing orderings}

\item{TO}{symbol indicating the start of a set of items ranked in a tie}

\item{TC}{symbol indicating the end of a set of items ranked in a tie}

\item{NAs}{symbol indicating missing ranked object. Can be any symbol in the original ordering matrix}
}
\value{
a ranking or a matrix of rankings:
\tabular{lll}{
R \tab  \tab ranking or matrix of rankings}
}
\description{
From ordering to rank. IMPORTANT: check which symbol denotes tied rankings in the X matrix
}
\examples{
data(APAred)
ord=rank2order(APAred) #transform rankings into orderings
ran=order2rank(ord) #transform the orderings into rankings

}
\author{
Antonio D'Ambrosio \email{antdambr@unina.it}
}
