\name{Penalty}
\alias{Penalty}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Auxiliary function
}
\description{
Assign a penalty to the branches of the FASTcons amd QuickCons algorithms 
}
\usage{
Penalty(CR, cij, indice)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{CR}{
candidate to be the consensus ranking
}
  \item{cij}{
Combined input matrix
}
  \item{indice}{other input called by other functions}

}

\references{
Amodio, S., D'Ambrosio, A. & Siciliano, R (2015). Accurate algorithms for identifying the median ranking when dealing with weak and partial rankings under the Kemeny axiomatic approach. European Journal of Operational Research. DOI: 10.1016/j.ejor.2015.08.048.
}
\author{
Antonio D'Ambrosio <antdambr@unina.it> and Sonia Amodio <sonia.amodio@unina.it>
}

