\name{PenaltyBB2}
\alias{PenaltyBB2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Auxiliary function
}
\description{
Assign a penalty to the branches of the BB algorithms
}
\usage{
PenaltyBB2(cij, candidate, ord)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{cij}{
Combined input matrix
}
  \item{candidate}{
ranking candidate to be the consensus ranking
}
  \item{ord}{other input called by other functions}
}

\references{
Emond, E. J., & Mason, D. W. (2002). A new rank correlation coefficient with application to the consensus ranking problem. Journal of Multi-Criteria Decision Analysis, 11(1), 17-28.
}
\author{
Antonio D'Ambrosio <antdambr@unina.it>
}

