% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reordering.R
\name{reordering}
\alias{reordering}
\title{Given a vector (or a matrix), returns an ordered vector (or a matrix with ordered vectors)}
\usage{
reordering(X)
}
\arguments{
\item{X}{a ranking, or a ranking data matrix}
}
\value{
a ranking in natural form
}
\description{
Given a ranking of M objects (or a matrix with M columns), it reduces it in "natural" form (i.e., with integers from 1 to M)
}
\author{
Antonio D'Ambrosio \email{antdambr@unina.it}
}
