% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReorderingBB.R
\name{ReorderingBB}
\alias{ReorderingBB}
\title{Auxiliary function}
\usage{
ReorderingBB(RR)
}
\arguments{
\item{RR}{A ranking}
}
\value{
A reordered ranking
}
\description{
It allows to reord the objects to be processed in the BB algorithms
}
\author{
Sonia Amodio \email{sonia.amodio@unina.it}
}
