\name{sports}
\alias{sports}
\docType{data}
\title{
sports data
}
\description{
130 students at the University of Illinois ranked seven sports according to their preference (Baseball, Football, Basketball, Tennis,  Cycling, Swimming, Jogging).
}
\usage{data(sports)}


\source{
Marden, J. I. (1996). Analyzing and modeling rank data. CRC Press.
}

\examples{
data(sports)
}
\keyword{datasets}
