% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findbranches.R
\name{findbranches}
\alias{findbranches}
\title{Auxiliary function}
\usage{
findbranches(R, ord, b, FULL = FALSE)
}
\arguments{
\item{R}{Candidate to be the consensus ranking}

\item{ord}{other input values recalled by other routines}

\item{b}{other input values recalled by other routines}

\item{FULL}{Default FULL=FALSE. If FULL=TRUE, the searching is limited to the space of complete rankings. TRUE=TRUE if the function is called by BBFULL algorithm.}
}
\value{
a candidate to be the median ranking
}
\description{
Find correct branches in the Branch-and-Bound algorithms
}
\author{
Antonio D'Ambrosio \email{antdambr@unina.it} and Sonia Amodio \email{sonia.amodio@unina.it}
}
