\name{findconsensusBB}
\alias{findconsensusBB}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Find a first ranking candidate to be the cinsensus ranking
}
\description{
Auxiliary function: it finds a first ranking candidate to be the consensus ranking by observing the Combined input matrix
}
\usage{
findconsensusBB(cij)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{cij}{
Combined input matrix
}
}

\value{
A candidate consensus ranking
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}

\author{
Antonio D'Ambrosio <antdambr@unina.it>
}
