% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kemenydesign.R
\name{kemenydesign}
\alias{kemenydesign}
\title{Auxiliary function}
\usage{
kemenydesign(X)
}
\arguments{
\item{X}{A N by M data matrix, in which there are N judges and M objects to be judged. Each row is a ranking of the objects represented by the columns.}
}
\value{
Design matrix
}
\description{
Define a design matrix to compute Kemeny distance
}
\references{
D'Ambrosio, A. (2008). Tree based methods for data editing and preference rankings. Unpublished PhD Thesis. Universita' degli Studi di Napoli Federico II.
}
\author{
Antonio D'Ambrosio \email{antdambr@unina.it}
}
