\name{Penalty}
\alias{Penalty}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Auxiliary function
}
\description{
Assign a penalty to the branches of the FASTcons amd QuickCons algorithms 
}
\usage{
Penalty(CR, cij, indice)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{CR}{
candidate to be the consensus ranking
}
  \item{cij}{
Combined input matrix
}
  \item{indice}{other input called by other functions}

}

\references{
Amodio, S., D'Ambrosio, A., & Siciliano, R. (2015). A Fast accurate heuristic algorithm for the consensus ranking problem. Submitted. Technical report available as arXiv preprint arXiv:1502.06498.
}
\author{
Antonio D'Ambrosio <antdambr@unina.it> and Sonia Amodio <sonia.amodio@unina.it>
}

