% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty.R
\name{Penalty}
\alias{Penalty}
\title{Auxiliary function}
\usage{
Penalty(CR, cij, indice)
}
\arguments{
\item{CR}{candidate to be the median ranking}

\item{cij}{combined input matrix}

\item{indice}{other input called by other functions}
}
\value{
a penalty value
}
\description{
Assign a penalty to the branches of the FASTcons amd QuickCons algorithms
}
\references{
Amodio, S., D'Ambrosio, A. and Siciliano, R. (2016). Accurate algorithms for identifying the median ranking when dealing with weak and partial rankings under the Kemeny axiomatic approach. European Journal of Operational Research, 249(2), 667-676.
}
\author{
Antonio D'Ambrosio \email{antdambr@unina.it} and Sonia Amodio \email{sonia.amodio@unina.it}
}
