% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PenaltyBB2.R
\name{PenaltyBB2}
\alias{PenaltyBB2}
\title{Auxiliary function}
\usage{
PenaltyBB2(cij, candidate, ord)
}
\arguments{
\item{cij}{combined input matrix}

\item{candidate}{candidate to be the median ranking}

\item{ord}{other input called by other functions}
}
\value{
computed penalty
}
\description{
Assign a penalty to the branches of the BB algorithms
}
\author{
Antonio D'Ambrosio \email{antdambr@unina.it}
}
