\name{reordering}
\alias{reordering}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Auxiliary function
}
\description{
Given a ranking of M objects (or a matrix with M columns), it reduces it in "natural" form (i.e., with integers from 1 to M)
}
\usage{
reordering(X)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
a ranking, or a ranking data matrix
}
}

\value{
a ranking in natural form
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}

\author{
Antonio D'Ambrosio <antdambr@unina.it> 
}

