\name{ReorderingBB}
\alias{ReorderingBB}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Auxiliary function
}
\description{
It allows to reord the objects to be processed in the BB algorithms
}
\usage{
ReorderingBB(RR)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{RR}{
A ranking
}
}

\value{
A reordered ranking
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}

\author{
Antonio D'Ambrosio <antdambr@unina.it> and Sonia Amodio <sonia.amodio@unina.it>
}

